requirements.txt
Django==5.1.3
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
drf-yasg==1.21.8
